SDPSymmetryReduction

Documentation for SDPSymmetryReduction.

SDPSymmetryReduction.admPartSubspace

SDPSymmetryReduction.blockDiagonalize

SDPSymmetryReduction.admPartSubspace — Method

admPartSubspace(C::Vector{T}, A::Matrix{T}, b::Vector{T}, verbose::Bool = false)

Compute the smallest admissible partion subspace for the SDP $\inf/\sup\{\dot(C,x), Ax = b, \mathrm\{Mat\}(x) \text{PSD/DNN}\}.$

SDPSymmetryReduction.blockDiagonalize — Function

blockDiagonalize(P::Partition, verbose = true; epsilon = 1e-8, complex = false)

Compute the block-diagonalization of the (Jordan)-algebra given by the Partition P. May fail if complex = false, since a real block diagonalization does not always exist.