SDPSymmetryReduction.jl

A symmetry reduction package for Julia.

Documentation for SDPSymmetryReduction.

This package provides functions to determine a symmetry reduction of an SDP numerically using the Jordan Reduction method.

It assumes that the problem is given in vectorized standard form

sup/inf     dot(C,x)
subject to  Ax = b,
            Mat(x) is positive semidefinite/doubly nonnegative,

where C and b are vectors and A is a matrix.

The function admPartSubspace finds an optimal admissible partition subspace for a given SDP. An SDP can be restricted to such a subspace without changing its optimum. The returned Partition-subspace can then be block-diagonalized using blockDiagonalize.

For details on the theory and the implemented algorithms, check out the reference linked in the repository.

Examples

The Theta'-function of Erdos-Renyi graphs

Symmetry reducting a strong relaxation of the quadratic assigment problem

General example

Documentation

SDPSymmetryReduction.Partition — Type

Partition

A partition subspace. P.n is the number of parts, and P.P an integer matrix defining the basis elements.

SDPSymmetryReduction.admPartSubspace — Method

admPartSubspace(C::Vector{T}, A::Matrix{T}, b::Vector{T}, verbose::Bool = false)

Returns the optimal admissible partition subspace for the SDP

$\inf\{C^Tx, Ax = b, \mathrm{Mat}(x) \succcurlyeq 0\}.$

This is done using a randomized Jordan-reduction algorithm, and it returns a Jordan algebra (closed under linear combinations and squaring). SDPs can be restricted to such a subspace without changing their optimal value.

Output

A Partition subspace P.

SDPSymmetryReduction.blockDiagonalize — Function

blockDiagonalize(P::Partition, verbose = true; epsilon = 1e-8, complex = false)

Determines a block-diagonalization of a (Jordan)-algebra given by a partition P using a randomized algorithm. blockDiagonalize(P) returns a real block-diagonalization blkd, if it exists, otherwise nothing.

blockDiagonalize(P; complex = true) returns the same, but with complex valued matrices, and should be used if no real block-diagonalization was found. To use the complex matrices practically, remember that a Hermitian matrix A is positive semidefinite iff [real(A) -imag(A); imag(A) real(A)] is positive semidefinite.

Output

blkd.blkSizes is an integer array of the sizes of the blocks.

blkd.blks is an array of length P.n containing arrays of (real/complex) matrices of sizes blkd.blkSizes. I.e. blkd.blks[i] is the image of the basis element P.P .== i.